requirements.txt
torch
torchvision
numpy
opencv-python
hydra-core
hydra-colorlog
ptflops
tqdm